+++ R/antibiogram.R
@@ -27,17 +27,18 @@
 # how to conduct AMR data analysis: https://msberends.github.io/AMR/   #
 # ==================================================================== #
 
-#' Generate Antibiogram: Traditional, Combined, Syndromic, or Weighted-Incidence Syndromic Combination (WISCA)
+#' Generate Traditional, Combination, Syndromic, or WISCA Antibiograms
 #'
-#' Generate an antibiogram, and communicate the results in plots or tables. These functions follow the logic of Klinker *et al.* and Barbieri *et al.* (see *Source*), and allow reporting in e.g. R Markdown and Quarto as well.
+#' Create detailed antibiograms with options for traditional, combination, syndromic, and Bayesian WISCA methods. Based on the approaches of Klinker *et al.*, Barbieri *et al.*, and the Bayesian WISCA model (Weighted-Incidence Syndromic Combination Antibiogram) by Bielicki *et al.*, this function provides flexible output formats including plots and tables, ideal for integration with R Markdown and Quarto reports.
 #' @param x a [data.frame] containing at least a column with microorganisms and columns with antibiotic results (class 'sir', see [as.sir()])
 #' @param antibiotics vector of any antibiotic name or code (will be evaluated with [as.ab()], column name of `x`, or (any combinations of) [antibiotic selectors][antibiotic_class_selectors] such as [aminoglycosides()] or [carbapenems()]. For combination antibiograms, this can also be set to values separated with `"+"`, such as "TZP+TOB" or "cipro + genta", given that columns resembling such antibiotics exist in `x`. See *Examples*.
-#' @param mo_transform a character to transform microorganism input - must be "name", "shortname", "gramstain", or one of the column names of the [microorganisms] data set: `r vector_or(colnames(microorganisms), sort = FALSE, quotes = TRUE)`. Can also be `NULL` to not transform the input.
-#' @param ab_transform a character to transform antibiotic input - must be one of the column names of the [antibiotics] data set: `r vector_or(colnames(antibiotics), sort = FALSE, quotes = TRUE)`. Can also be `NULL` to not transform the input.
+#' @param mo_transform a character to transform microorganism input - must be `"name"`, `"shortname"` (default), `"gramstain"`, or one of the column names of the [microorganisms] data set: `r vector_or(colnames(microorganisms), sort = FALSE, quotes = TRUE)`. Can also be `NULL` to not transform the input.
+#' @param ab_transform a character to transform antibiotic input - must be one of the column names of the [antibiotics] data set (defaults to "name"): `r vector_or(colnames(antibiotics), sort = FALSE, quotes = TRUE)`. Can also be `NULL` to not transform the input.
 #' @param syndromic_group a column name of `x`, or values calculated to split rows of `x`, e.g. by using [ifelse()] or [`case_when()`][dplyr::case_when()]. See *Examples*.
 #' @param add_total_n a [logical] to indicate whether total available numbers per pathogen should be added to the table (default is `TRUE`). This will add the lowest and highest number of available isolate per antibiotic (e.g, if for *E. coli* 200 isolates are available for ciprofloxacin and 150 for amoxicillin, the returned number will be "150-200").
 #' @param only_all_tested (for combination antibiograms): a [logical] to indicate that isolates must be tested for all antibiotics, see *Details*
-#' @param digits number of digits to use for rounding
+#' @param digits number of digits to use for rounding the susceptibility percentage
+#' @param formatting_type numeric value (1–12) indicating how the 'cells' of the antibiogram table should be formatted. See *Details* > *Formatting Type* for a list of options.
 #' @param col_mo column name of the names or codes of the microorganisms (see [as.mo()]) - the default is the first column of class [`mo`]. Values will be coerced using [as.mo()].
 #' @param language language to translate text, which defaults to the system language (see [get_AMR_locale()])
 #' @param minimum the minimum allowed number of available (tested) isolates. Any isolate count lower than `minimum` will return `NA` with a warning. The default number of `30` isolates is advised by the Clinical and Laboratory Standards Institute (CLSI) as best practice, see *Source*.
@@ -47,14 +48,35 @@
 #' @param object an [antibiogram()] object
 #' @param ... when used in [R Markdown or Quarto][knitr::kable()]: arguments passed on to [knitr::kable()] (otherwise, has no use)
 #' @details This function returns a table with values between 0 and 100 for *susceptibility*, not resistance.
-#'
+#' 
 #' **Remember that you should filter your data to let it contain only first isolates!** This is needed to exclude duplicates and to reduce selection bias. Use [first_isolate()] to determine them in your data set with one of the four available algorithms.
-#'
-#' All types of antibiograms as listed below can be plotted (using [ggplot2::autoplot()] or base \R [plot()]/[barplot()]). The `antibiogram` object can also be used directly in R Markdown / Quarto (i.e., `knitr`) for reports. In this case, [knitr::kable()] will be applied automatically and microorganism names will even be printed in italics at default (see argument `italicise`). You can also use functions from specific 'table reporting' packages to transform the output of [antibiogram()] to your needs, e.g. with `flextable::as_flextable()` or `gt::gt()`.
+#' 
+#' ### Formatting Type
+#' 
+#' The formatting of the 'cells' of the table can be set with the argument `formatting_type`. In these examples, `5` is the susceptibility percentage, `15` the numerator, and `300` the denominator:
+#' 
+#' 1. 5
+#' 2. 15
+#' 3. 300
+#' 4. 15/300
+#' 5. 5 (300)
+#' 6. 5% (300)
+#' 7. 5 (N=300)
+#' 8. 5% (N=300)
+#' 9. 5 (15/300)
+#' 10. 5% (15/300)
+#' 11. 5 (N=15/300)
+#' 12. 5% (N=15/300)
+#' 
+#' The default is `10`, which can be set globally with the [package option][AMR-options] [`AMR_antibiogram_formatting_type`][AMR-options], e.g. `options(AMR_antibiogram_formatting_type = 5)`.
+#' 
+#' Set `digits` (defaults to `0`) to alter the rounding of the susceptibility percentage.
 #'
 #' ### Antibiogram Types
 #'
-#' There are four antibiogram types, as proposed by Klinker *et al.* (2021, \doi{10.1177/20499361211011373}), and they are all supported by [antibiogram()]:
+#' There are four antibiogram types, as summarised by Klinker *et al.* (2021, \doi{10.1177/20499361211011373}), and they are all supported by [antibiogram()]. Use WISCA whenever possible, since it provides precise coverage estimates by accounting for pathogen incidence and antimicrobial susceptibility. See the section *Why Use WISCA?* on this page.
+#' 
+#' The four antibiogram types:
 #'
 #' 1. **Traditional Antibiogram**
 #'
@@ -91,6 +113,8 @@
 #'    ```
 #'
 #' 4. **Weighted-Incidence Syndromic Combination Antibiogram (WISCA)**
+#' 
+#'    WISCA enhances empirical antibiotic selection by weighting the incidence of pathogens in specific clinical syndromes and combining them with their susceptibility data. It provides an estimation of regimen coverage by aggregating pathogen incidences and susceptibilities across potential causative organisms. See also the section *Why Use WISCA?* on this page.
 #'
 #'    Case example: Susceptibility of *Pseudomonas aeruginosa* to TZP among respiratory specimens (obtained among ICU patients only) for male patients age >=65 years with heart failure
 #'
@@ -106,8 +130,12 @@
 #'                                             .$condition == "Heart Disease",
 #'                                           "Study Group", "Control Group"))
 #'    ```
+#'    
+#'    WISCA uses a sophisticated Bayesian decision model to combine both local and pooled antimicrobial resistance data. This approach not only evaluates local patterns but can also draw on multi-centre datasets to improve regimen accuracy, even in low-incidence infections like paediatric bloodstream infections (BSIs).
 #'
-#' Note that for combination antibiograms, it is important to realise that susceptibility can be calculated in two ways, which can be set with the `only_all_tested` argument (default is `FALSE`). See this example for two antibiotics, Drug A and Drug B, about how [antibiogram()] works to calculate the %SI:
+#' ### Inclusion in Combination Antibiogram and Syndromic Antibiogram
+#'
+#' Note that for types 2 and 3 (Combination Antibiogram and Syndromic Antibiogram), it is important to realise that susceptibility can be calculated in two ways, which can be set with the `only_all_tested` argument (default is `FALSE`). See this example for two antibiotics, Drug A and Drug B, about how [antibiogram()] works to calculate the %SI:
 #'
 #' ```
 #' --------------------------------------------------------------------
@@ -127,8 +155,28 @@
 #'   <NA>      <NA>        -            -            -            -
 #' --------------------------------------------------------------------
 #' ```
+#' 
+#' ### Plotting
+#' 
+#' All types of antibiograms as listed above can be plotted (using [ggplot2::autoplot()] or base \R's [plot()] and [barplot()]).
+#' 
+#' THe outcome of [antibiogram()] can also be used directly in R Markdown / Quarto (i.e., `knitr`) for reports. In this case, [knitr::kable()] will be applied automatically and microorganism names will even be printed in italics at default (see argument `italicise`).
+#' 
+#' You can also use functions from specific 'table reporting' packages to transform the output of [antibiogram()] to your needs, e.g. with `flextable::as_flextable()` or `gt::gt()`.
 #'
+#' @section Why Use WISCA?:
+#' WISCA is a powerful tool for guiding empirical antibiotic therapy because it provides precise coverage estimates by accounting for pathogen incidence and antimicrobial susceptibility. This is particularly important in empirical treatment, where the causative pathogen is often unknown at the outset. Traditional antibiograms do not reflect the weighted likelihood of specific pathogens based on clinical syndromes, which can lead to suboptimal treatment choices. 
+#' 
+#' The Bayesian WISCA, as described by Bielicki *et al.* (2016), improves on earlier methods by handling uncertainties common in smaller datasets, such as low-incidence infections. This method offers a significant advantage by: 
+#' 
+#' 1. Pooling Data from Multiple Sources:\cr WISCA uses pooled data from multiple hospitals or surveillance sources to overcome limitations of small sample sizes at individual institutions, allowing for more confident selection of narrow-spectrum antibiotics or combinations. 
+#' 2. Bayesian Framework:\cr The Bayesian decision tree model accounts for both local data and prior knowledge (such as inherent resistance patterns) to estimate regimen coverage. It allows for a more precise estimation of coverage, even in cases where susceptibility data is missing or incomplete. 
+#' 3. Incorporating Pathogen and Regimen Uncertainty:\cr WISCA allows clinicians to see the likelihood that an empirical regimen will be effective against all relevant pathogens, taking into account uncertainties related to both pathogen prevalence and antimicrobial resistance. This leads to better-informed, data-driven clinical decisions. 
+#' 4. Scenarios for Optimising Treatment:\cr For hospitals or settings with low-incidence infections, WISCA helps determine whether local data is sufficient or if pooling with external data is necessary. It also identifies statistically significant differences or similarities between antibiotic regimens, enabling clinicians to choose optimal therapies with greater confidence. 
+#' 
+#' WISCA is essential in optimising empirical treatment by shifting away from broad-spectrum antibiotics, which are often overused in empirical settings. By offering precise estimates based on syndromic patterns and pooled data, WISCA supports antimicrobial stewardship by guiding more targeted therapy, reducing unnecessary broad-spectrum use, and combating the rise of antimicrobial resistance.
 #' @source
+#' * Bielicki JA *et al.* (2016). **Selecting appropriate empirical antibiotic regimens for paediatric bloodstream infections: application of a Bayesian decision model to local and pooled antimicrobial resistance surveillance data** *Journal of Antimicrobial Chemotherapy* 71(3); \doi{10.1093/jac/dkv397}
 #' * Klinker KP *et al.* (2021). **Antimicrobial stewardship and antibiograms: importance of moving beyond traditional antibiograms**. *Therapeutic Advances in Infectious Disease*, May 5;8:20499361211011373; \doi{10.1177/20499361211011373}
 #' * Barbieri E *et al.* (2021). **Development of a Weighted-Incidence Syndromic Combination Antibiogram (WISCA) to guide the choice of the empiric antibiotic treatment for urinary tract infection in paediatric patients: a Bayesian approach** *Antimicrobial Resistance & Infection Control* May 1;10(1):74; \doi{10.1186/s13756-021-00939-2}
 #' * **M39 Analysis and Presentation of Cumulative Antimicrobial Susceptibility Test Data, 5th Edition**, 2022, *Clinical and Laboratory Standards Institute (CLSI)*. <https://clsi.org/standards/products/microbiology/documents/m39/>.
@@ -253,11 +301,12 @@
 antibiogram <- function(x,
                         antibiotics = where(is.sir),
                         mo_transform = "shortname",
-                        ab_transform = NULL,
+                        ab_transform = "name",
                         syndromic_group = NULL,
-                        add_total_n = TRUE,
+                        add_total_n = FALSE,
                         only_all_tested = FALSE,
                         digits = 0,
+                        formatting_type = getOption("AMR_antibiogram_formatting_type", 10),
                         col_mo = NULL,
                         language = get_AMR_locale(),
                         minimum = 30,
@@ -271,6 +320,7 @@ antibiogram <- function(x,
   meet_criteria(add_total_n, allow_class = "logical", has_length = 1)
   meet_criteria(only_all_tested, allow_class = "logical", has_length = 1)
   meet_criteria(digits, allow_class = c("numeric", "integer"), has_length = 1, is_finite = TRUE)
+  meet_criteria(formatting_type, allow_class = c("numeric", "integer"), has_length = 1, is_in = c(1:12))
   meet_criteria(col_mo, allow_class = "character", has_length = 1, allow_NULL = TRUE, is_in = colnames(x))
   language <- validate_language(language)
   meet_criteria(minimum, allow_class = c("numeric", "integer"), has_length = 1, is_positive_or_zero = TRUE, is_finite = TRUE)
@@ -388,7 +438,7 @@ antibiogram <- function(x,
   counts <- out
 
   if (isTRUE(combine_SI)) {
-    out$numerator <- out$S + out$I
+    out$numerator <- out$S + out$I + out$SDD
   } else {
     out$numerator <- out$S
   }
@@ -413,8 +463,36 @@ antibiogram <- function(x,
       pm_group_by(mo, ab)
   }
   
-  out <- out %pm>%
-    pm_summarise(SI = numerator / total)
+  # formatting type:
+  # 1. 5
+  # 2. 15
+  # 3. 300
+  # 4. 15/300
+  # 5. 5 (300)
+  # 6. 5% (300)
+  # 7. 5 (N=300)
+  # 8. 5% (N=300)
+  # 9. 5 (15/300)
+  # 10. 5% (15/300)
+  # 11. 5 (N=15/300)
+  # 12. 5% (N=15/300)
+  out_numeric <- out %pm>%
+    pm_summarise(percentage = numerator / total,
+                 numerator = numerator,
+                 total = total)
+  out$digits <- digits # since pm_sumarise() cannot work with an object outside the current frame
+  if (formatting_type == 1) out <- out %pm>% pm_summarise(out_value = round((numerator / total) * 100, digits = digits))
+  if (formatting_type == 2) out <- out %pm>% pm_summarise(out_value = numerator)
+  if (formatting_type == 3) out <- out %pm>% pm_summarise(out_value = total)
+  if (formatting_type == 4) out <- out %pm>% pm_summarise(out_value = paste0(numerator, "/", total))
+  if (formatting_type == 5) out <- out %pm>% pm_summarise(out_value = paste0(round((numerator / total) * 100, digits = digits), " (", total, ")"))
+  if (formatting_type == 6) out <- out %pm>% pm_summarise(out_value = paste0(round((numerator / total) * 100, digits = digits), "% (", total, ")"))
+  if (formatting_type == 7) out <- out %pm>% pm_summarise(out_value = paste0(round((numerator / total) * 100, digits = digits), " (N=", total, ")"))
+  if (formatting_type == 8) out <- out %pm>% pm_summarise(out_value = paste0(round((numerator / total) * 100, digits = digits), "% (N=", total, ")"))
+  if (formatting_type == 9) out <- out %pm>% pm_summarise(out_value = paste0(round((numerator / total) * 100, digits = digits), " (", numerator, "/", total, ")"))
+  if (formatting_type == 10) out <- out %pm>% pm_summarise(out_value = paste0(round((numerator / total) * 100, digits = digits), "% (", numerator, "/", total, ")"))
+  if (formatting_type == 11) out <- out %pm>% pm_summarise(out_value = paste0(round((numerator / total) * 100, digits = digits), " (N=", numerator, "/", total, ")"))
+  if (formatting_type == 12) out <- out %pm>% pm_summarise(out_value = paste0(round((numerator / total) * 100, digits = digits), "% (N=", numerator, "/", total, ")"))
   
   # transform names of antibiotics
   ab_naming_function <- function(x, t, l, s) {
@@ -437,15 +515,15 @@ antibiogram <- function(x,
     out
   }
   out$ab <- ab_naming_function(out$ab, t = ab_transform, l = language, s = sep)
-
+  out_numeric$ab <- ab_naming_function(out_numeric$ab, t = ab_transform, l = language, s = sep)
+  
   # transform long to wide
-  long_to_wide <- function(object, digs) {
-    object$SI <- round(object$SI * 100, digits = digs)
+  long_to_wide <- function(object) {
     object <- object %pm>%
       # an unclassed data.frame is required for stats::reshape()
       as.data.frame(stringsAsFactors = FALSE) %pm>%
-      stats::reshape(direction = "wide", idvar = "mo", timevar = "ab", v.names = "SI")
-    colnames(object) <- gsub("^SI?[.]", "", colnames(object))
+      stats::reshape(direction = "wide", idvar = "mo", timevar = "ab", v.names = "out_value")
+    colnames(object) <- gsub("^out_value?[.]", "", colnames(object))
     return(object)
   }
 
@@ -453,18 +531,17 @@ antibiogram <- function(x,
   attr(out, "pm_groups") <- NULL
   attr(out, "groups") <- NULL
   class(out) <- class(out)[!class(out) %in% c("grouped_df", "grouped_data")]
-  long <- out
 
   if (isTRUE(has_syndromic_group)) {
     grps <- unique(out$syndromic_group)
     for (i in seq_len(length(grps))) {
       grp <- grps[i]
       if (i == 1) {
-        new_df <- long_to_wide(out[which(out$syndromic_group == grp), , drop = FALSE], digs = digits)
+        new_df <- long_to_wide(out[which(out$syndromic_group == grp), , drop = FALSE])
       } else {
         new_df <- rbind_AMR(
           new_df,
-          long_to_wide(out[which(out$syndromic_group == grp), , drop = FALSE], digs = digits)
+          long_to_wide(out[which(out$syndromic_group == grp), , drop = FALSE])
         )
       }
     }
@@ -474,7 +551,7 @@ antibiogram <- function(x,
     new_df <- new_df[, c("syndromic_group", "mo", sort(colnames(new_df)[!colnames(new_df) %in% c("syndromic_group", "mo")])), drop = FALSE]
     colnames(new_df)[1:2] <- translate_AMR(c("Syndromic Group", "Pathogen"), language = language)
   } else {
-    new_df <- long_to_wide(out, digs = digits)
+    new_df <- long_to_wide(out)
     # sort rows
     new_df <- new_df %pm>% pm_arrange(mo)
     # sort columns
@@ -514,19 +591,14 @@ antibiogram <- function(x,
   rownames(out) <- NULL
   structure(out,
     has_syndromic_group = has_syndromic_group,
-    long = long,
+    out_numeric = out_numeric,
     combine_SI = combine_SI
   )
 }
 
 # will be exported in R/zzz.R
 tbl_sum.antibiogram <- function(x, ...) {
-  if (isTRUE(base::l10n_info()$`UTF-8`)) {
-    cross <- "\u00d7"
-  } else {
-    cross <- "x"
-  }
-  dims <- paste(format(NROW(x), big.mark = ","), cross, format(NCOL(x), big.mark = ","))
+  dims <- paste(format(NROW(x), big.mark = ","), AMR_env$cross_icon, format(NCOL(x), big.mark = ","))
   names(dims) <- "An Antibiogram"
   dims
 }
@@ -545,7 +617,7 @@ tbl_format_footer.antibiogram <- function(x, ...) {
 #' @export
 #' @rdname antibiogram
 plot.antibiogram <- function(x, ...) {
-  df <- attributes(x)$long
+  df <- attributes(x)$out_numeric
   if ("syndromic_group" %in% colnames(df)) {
     # barplot in base R does not support facets - paste columns together
     df$mo <- paste(df$mo, "-", df$syndromic_group)
@@ -561,7 +633,7 @@ plot.antibiogram <- function(x, ...) {
     df_sub <- df[df$mo == mo, , drop = FALSE]
 
     barplot(
-      height = df_sub$SI * 100,
+      height = df_sub$percentage * 100,
       xlab = NULL,
       ylab = ifelse(isTRUE(attributes(x)$combine_SI), "%SI", "%S"),
       names.arg = df_sub$ab,
@@ -584,12 +656,12 @@ barplot.antibiogram <- function(height, ...) {
 #' @rdname antibiogram
 # will be exported using s3_register() in R/zzz.R
 autoplot.antibiogram <- function(object, ...) {
-  df <- attributes(object)$long
+  df <- attributes(object)$out_numeric
   ggplot2::ggplot(df) +
     ggplot2::geom_col(
       ggplot2::aes(
         x = ab,
-        y = SI * 100,
+        y = percentage * 100,
         fill = if ("syndromic_group" %in% colnames(df)) {
           syndromic_group
         } else {